refactor: substitui Postgres/JDBC por BigQuery table functions

+++ src/main/java/com/rmfarma/pharmahub/api/dto/response/queries/StockMetricsDTO.java
@@ -6,7 +6,7 @@ public record StockMetricsDTO(
         String cnpj,
         String grupoEconomico,
         BigDecimal totalCustoEstoque,
-        Long totalItensAltaRotatividade
+        BigDecimal totalItensAltaRotatividade
 ) {
 }
